ID and TDNestCnt counters are originally BYTE types
git-svn-id: https://svn.aros.org/svn/aros/trunk@50882 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/arm-native/kernel/tls.h
@@ -7,8 +7,8 @@ typedef struct tls
     void *              *KernelBase;    /* Base of kernel.resource */
     struct Task         *ThisTask;      /* Currently running task on this core */
     ULONG               ScheduleFlags;
-    LONG                IDNestCnt;
-    LONG                TDNestCnt;
+    BYTE                IDNestCnt;
+    BYTE                TDNestCnt;
 } tls_t;
 
 #define TLSSF_Quantum   (1 << 0)
